Move database-client.mjs into db directory.
